decomp: TubePlate::AllocateVerts cast capacity() to int
bin/find-inlining-gaps flagged MakeString<int, unsigned short> vs
target's MakeString<int, int>. verts.capacity() returns unsigned short
(vector<Vert, unsigned short>); cast resolves the bl mismatch.

99.4 -> 99.5%; residual is r25↔r27 callee-saved cascade.

+++ src/system/bandobj/NoteTube.cpp
@@ -425,7 +425,7 @@ void TubePlate::AllocateVerts(int num, bool warn) {
             MILO_WARN(
                 "TubePlate: Reallocating verts from %d to %d; please alert HUD/Track owner",
                 cap,
-                verts.capacity()
+                (int)verts.capacity()
             );
     }
     verts.resize(newsize, true);
